myaimediamgr_project/myaimediamgr-backend/requirements.txt
Flask==3.1.1
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
psycopg2-binary==2.9.9

google-generativeai==0.7.2
Pillow==10.3.0
numpy==1.26.4
moviepy==1.0.3
imageio==2.34.1
imageio-ffmpeg==0.4.9
